add a indicator while load app store view
add a indicator while load app store view

+++ vCode/md5Encryptor.m
@@ -16,7 +16,7 @@
 {
     const char *cStr = [str UTF8String];
     unsigned char result[16];
-    CC_MD5(cStr, strlen(cStr), result); // This is the md5 call
+    CC_MD5(cStr, (unsigned int)strlen(cStr), result); // This is the md5 call
     return [NSString stringWithFormat:
             @"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
             result[0], result[1], result[2], result[3],
